[chore] ut 위치 수
